Couple of changes in the template logic:
Check the length of the remaining number of results and pass a boolean
value to the template to decide whether to show the next link or not.

In the results template, fix the logic of showing the previous link. It
was showing it only if the current page > 1, while the page count is
starting at 0, so it should have been page >= 1. Also, the previous
link should point to page -1 rather than page.

And show the next link only if next_page = True